abcd split into several files.  This version compiles and runs all examples
Solution
